tests/requirements.txt
# App01 TalentAI 功能测试依赖
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-html>=4.0.0
pytest-xdist>=3.5.0
pytest-cov>=4.0.0
httpx>=0.27.0
requests>=2.31.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
